analysis/kusto/requirements.txt
azure-kusto-data>=4.0.0
pandas>=2.0.0
altair>=5.0.0
nbstripout>=0.7.0
